Create DeckManager class for CommandSupport
This allows proper stubbing and spying of Deck-related functions and the extraction of further commands from TournamentManager. This also consolidates deck and discordDeck into one file. This is a bit of a monolithic commit because the way everything was intertwined made it hard to be atomic, sorry. The test for the deck command has been expanded.
