+++ test/Command.unit.ts
@@ -4,6 +4,7 @@ import sinon, { SinonSandbox } from "sinon";
 import sinonChai from "sinon-chai";
 import sinonTest from "sinon-test";
 import { Command, CommandDefinition, CommandSupport } from "../src/Command";
+import { DeckManager } from "../src/deck";
 import { DiscordInterface } from "../src/discord/interface";
 import { OrganiserRoleProvider } from "../src/role/organiser";
 import { UserError } from "../src/util/errors";
@@ -24,7 +25,8 @@ describe("Command class", function () {
 		// UNUSED
 		database: new DatabaseWrapperMock(),
 		challonge: new WebsiteInterface(new WebsiteWrapperMock()),
-		scores: new Map()
+		scores: new Map(),
+		decks: new DeckManager([])
 	};
 	const msg = new Message({ id: "007", channel_id: "foo", author: { id: "0000" } }, new Client("mock"));
 	const testCommand: CommandDefinition = {
