+++ test/mocks/tournament.ts
@@ -1,6 +1,4 @@
-import { Deck } from "ydeck";
 import { DatabasePlayer } from "../../src/database/interface";
-import { getDeck } from "../../src/deck/deck";
 import { TournamentInterface } from "../../src/TournamentManager";
 
 export class TournamentMock implements TournamentInterface {
@@ -84,13 +82,6 @@ export class TournamentMock implements TournamentInterface {
 
 	public async getConfirmed(): Promise<DatabasePlayer[]> {
 		return [];
-	}
-
-	public async getPlayerDeck(): Promise<Deck> {
-		// ABC test deck from YDeck test suites
-		return await getDeck(
-			"ydke://5m3qBeZt6gV9+McCffjHAn34xwK8beUDvG3lA7xt5QMfX5ICWvTJAVr0yQFa9MkBrDOdBKwznQSsM50Ey/UzAMv1MwDL9TMAdAxQBQ6wYAKvI94AryPeAK8j3gCmm/QBWXtjBOMavwDjGr8A4xq/AD6kcQGeE8oEnhPKBJ4TygSlLfUDpS31A6Ut9QMiSJkAIkiZACJImQCANVMDgDVTAw==!FtIXALVcnwC1XJ8AiBF2A4gRdgNLTV4Elt0IAMf4TQHCT0EAvw5JAqSaKwD5UX8EweoDA2LO9ATaI+sD!H1+SAg==!"
-		);
 	}
 
 	public async dropPlayer(): Promise<void> {
